% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cell_visualizer}
\alias{cell_visualizer}
\title{generates the html document}
\usage{
cell_visualizer(taxid = "9606", sls = list("SL0073"), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
generates the html document
}
