% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCell.R
\name{drawCell}
\alias{drawCell}
\title{Creates a cell picture}
\usage{
drawCell(
  organism_identifier,
  sl_ids,
  inputId = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  colour_sl = NULL
)
}
\arguments{
\item{organism_identifier}{Taxonomy id of the organism of interest.
Can be found at https://www.ncbi.nlm.nih.gov/taxonomy/}

\item{sl_ids}{Id of the subcellular location, can be found at
http://current.geneontology.org/ontology/external2go/uniprotkb_sl2go}

\item{inputId}{input id of cell' defaults to `NULL`}

\item{width}{width of cell; defaults to `NULL`}

\item{height}{height of cell; defaults to `NULL`}

\item{elementId}{id of cell, defaults to `NULL`}

\item{colour_sl}{Highlighted color of the subcelullar components}
}
\value{
Generates a PNG of the requested cell by the user with highlighted
subcellular elements.
}
\description{
Creates a cell picture
}
\examples{
drawCell(organism_identifier = '58334', sl_ids = '0049', color = 'lightblue')

}
