% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edit_html}
\alias{edit_html}
\title{changes the size of the cell}
\usage{
edit_html(size = 2000)
}
\arguments{
\item{size}{size of the highlited element}
}
\value{
html modified with size
}
\description{
changes the size of the cell
}
