% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{edit_jscode}
\alias{edit_jscode}
\title{Changes the color of the cell}
\usage{
edit_jscode(color = "blue")
}
\arguments{
\item{color}{Color of the highlited element}
}
\value{
jscode modified with the color selected
}
\description{
Changes the color of the cell
}
