% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_custom_element.R
\name{generate_custom_element}
\alias{generate_custom_element}
\title{Creates custom_element.txt}
\usage{
generate_custom_element(...)
}
\arguments{
\item{...}{}
}
\value{
Creates custom_element.txt inside inst/snippet/ with the path to the
JScode, since it will be different to each user.
}
\description{
Creates custom_element.txt
}
