% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_html}
\alias{make_html}
\title{Generates the HTML document with the cell picture}
\usage{
make_html(
  taxonomy_id,
  sl_ids,
  templates = system.file("snippets/templates.txt", package = "drawCell"),
  custom_element = system.file("snippets/custom_element.txt", package = "drawCell")
)
}
\arguments{
\item{taxonomy_id}{taxonomy id}

\item{sl_ids}{subcelullar ids}

\item{templates}{snippets folders template.txt}

\item{custom_element}{snippets folders custom.txt}
}
\value{
html document
}
\description{
Generates the HTML document with the cell picture
}
