% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{restore_jscode}
\alias{restore_jscode}
\title{Restores the javascript code}
\usage{
restore_jscode(color = "blue")
}
\arguments{
\item{color}{Color of the highlited element}
}
\value{
jscode restored from the color selected
}
\description{
Restores the javascript code
}
