% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCell.R
\name{drawCell}
\alias{drawCell}
\title{Creates a cell picture}
\usage{
drawCell(taxonomy_id, sl_ids, color = "blue", size = 2000, delay = 1)
}
\arguments{
\item{taxonomy_id}{Taxonomy id of the organism of interest. Can be found at
https://www.ncbi.nlm.nih.gov/taxonomy/}

\item{sl_ids}{Id of the subcellular location, can be found at
http://current.geneontology.org/ontology/external2go/uniprotkb_sl2go}

\item{color}{Highlited color of the subcelullar components.}

\item{size}{Size of the cell, around 2000 should work well.}

\item{delay}{Time needed to make picture of the HTML (webshot). If the
picture is blank, try increasing the delay since it might take longer to
render the HTML. (Units in seconds)}
}
\value{
Generates a PNG of the requested cell by the user with highligthed
subcellular elements.
}
\description{
Creates a cell picture
}
\examples{
drawCell(taxonomy_id = '58334',
sl_ids = '0049',
size = 2000,
color = 'lightblue')
}
